requirements.txt
py-cord==2.6.1
pymongo==4.12.0
pytz==2025.2
python-dotenv==1.2.1